+++ PlatePath.API/Data/Models/Recipes/Ingredient.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace PlatePath.API.Data.Models.Recipe
+namespace PlatePath.API.Data.Models.Recipes
 {
     public class Ingredient
     {
